requirements.txt
beautifulsoup4==4.12.3
mechanize==0.4.10
python-dotenv==1.0.1
requests==2.32.3
